+++ tailbone/compute_engine/turn/__init__.py
@@ -18,6 +18,7 @@ from tailbone import config
 from tailbone import DEBUG
 from tailbone import PREFIX
 from tailbone.compute_engine import TailboneCEInstance
+from tailbone.compute_engine import STARTUP_SCRIPT_BASE
 
 import binascii
 from hashlib import sha1
@@ -27,6 +28,8 @@ import time
 import webapp2
 
 from google.appengine.api import lib_config
+from google.appengine.ext import ndb
+
 
 _config = lib_config.register("tailbone_turn", {
                               "SECRET": "notasecret",
@@ -41,40 +44,29 @@ class TailboneTurnInstance(TailboneCEInstance):
       "items": [
         {
           "key": "startup-script",
-          "value": """#!/bin/bash
-
-# install deps
-apt-get install -y build-essential python-dev
-
-# load reporter
-curl -O http://psutil.googlecode.com/files/psutil-0.6.1.tar.gz
-tar xvfz psutil-0.6.1.tar.gz
-cd psutil-0.6.1
-python setup.py install
-cd ..
-rm -rf psutil-0.6.1
-rm psutil-0.6.1.tar.gz
-curl -O https://raw.github.com/dataarts/tailbone/mesh/tailbone/compute_engine/load_reporter.py
-python load_reporter.py &
-
+          "value": STARTUP_SCRIPT_BASE + """
 # load turnserver
 curl -O http://rfc5766-turn-server.googlecode.com/files/turnserver-1.8.7.0-binary-linux-wheezy-ubuntu-mint-x86-64bits.tar.gz
 tar xvfz turnserver-1.8.7.0-binary-linux-wheezy-ubuntu-mint-x86-64bits.tar.gz
 dpkg -i rfc5766-turn-server_1.8.7.0-1_amd64.deb
 apt-get -f install
-turnserver --use-auth-secret -v -a -X -f --static-auth-secret notasecret -r localhost -r appspot.com
+turnserver --use-auth-secret -v -a -X -f --static-auth-secret %s %s
  
-""",
+""" % (_config.SECRET, " ".join(["-r " + str(d) in _config.RESTRICTED_DOMAINS]),
         },
       ],
     }
   })
 
+  secret = ndb.StringProperty(default=_config.SECRET)
 
-def credentials(username):
+
+def credentials(username, secret=None):
   timestamp = str(time.mktime(time.gmtime())).split('.')[0]
   username = "{}:{}".format(username, timestamp)
-  password = hmac.new(_config.SECRET, username, sha1)
+  if not secret:
+    secret = _config.SECRET
+  password = hmac.new(secret, username, sha1)
   password = binascii.b2a_base64(password.digest())[:-1]
   return username, password
 
@@ -93,5 +85,5 @@ class TurnHandler(BaseHandler):
     }
 
 app = webapp2.WSGIApplication([
-  (r"{}turn/?.*".format(PREFIX), TurnHandler),
+  (r"{}/turn/?.*".format(PREFIX), TurnHandler),
 ], debug=DEBUG)