+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+Version 1.0.0
+-----------
+
+  * Use of tailbone as a submodule
+  * New libraries for compute engine
+  * appengine_config.py for configuration
+
+
 Version 0.1.0
 -----------